Implement and test cellularSockFcntl() and, as part of this, test/fix non-blocking sockets behaviour.  In addition, SARA-R4 can take 30 seconds to close a TCP socket while it waits strictly for the ack to the ack to the ack, which kinda breaks non-blocking, so if non-blocking is set and we're on SARA-R4, add the parameter to the AT+SOCL command which causes the module to return immediately and send a URC when the socket is actually closed.  This necessitated adding a CLOSING state to the socket states so that the sendx()/receivex() functions can return immediately with an error if the socket is in that half-way-house state.
In the cellular connect API, remove AT+CREG=1 and AT+CGREG=1 since we don't support non-LTE RANs.

All tests pass but there is a sign of corruption occurring somewhere: I have seen a very innocent RAT change test fail because one of the array values seemed to be stomped on and, sometimes, we never get any data at all back from any of the sockets tests which transfer data.  This begins working again some time later without any change in the code.  This happens in one run on both of the echo servers we use, so it's not the server. It's not SARA-R4 'cos that's being rebooted frequently anyway.  Could be the Nutaq box?  There was a multi-threading bug which I came across in ESP-IDF v3 and moved to ESP-IDF v4 to avoid; we're on ESP-IDF v3.3.1 here, not sure if it was fixed in that version.  Need to watch points.

That's all the implementation I'm going to do on the sock/ctrl/port implementations for now: the TCP server related functionality can't be written until there is some way to test it (cellular networks and our u-blox LAN won't allow incoming TCP connections).

Next: try stuffing all this under AWS SDK.