Fixed the two bugs mentioned in commit #5aa20fa30: the strange value in the RAT change test was a bug in the test code of setGetRatRank() in cellular_ctrl_test.c (the test uses rand() and so needed two consecutive runs to show the problem) while the "not getting any data back at all" problem came about after a run of the maxNumSockets() test which, though it passed, would cause subsequent runs of any data transfer test to fail due to a bug in the way that statically allocated containers were added to the list at initialisation and another bug in the way that sockets were created.
With the bugs fixed the tests have been re-ordered so that the max sockets, socket options and non-blocking tests are naturally run between the data transfer tests so as to catch any similar problems in future.  All tests now 100% pass on two consecutive complete runs.

Also the locking of gMutexCallbacks in the URC wasn't wide enough, fixed now.