Switch power-saving off in the modem configuration function since there is no API for it yet and MNO profile 100, which we must use in Europe, switches it on, meaning that the modem appears to crash under some circumstances (e.g. after closing a TCP socket).
For SARA-R4, always ask for asynchronous socket close for a connected TCP socket, even if the socket is blocking, as it takes more than 35 seconds to get the ack for the ack for the ack for the ack and that's just plain too long.

Move the power-on and aliveness tests slightly later in the ctrl test sequence to prevent them failing as a consequence of a failure in the previous test run.

Call the run_unit_tests_and_detect_outcome.py script in the right location.