- UART implementation now uses the one HW timer we already claim as a timeout mechanism and hence doesn't have to do per-character interrupts; timer stuff moved into cellular_port_private.c/.h since it growded.  So far so good. - discovered that the Segger Embedded Studio compiler is set to use minimum-width enums, kind of a good catch as it breaks my usage of enums also as size values. Now fixed in the code by adding a 0x7FFFFFFF value to the relevant enums. - throw away the guff that SARA-R5 spews out at boot 'cos it gets in the way. - remove the cellularPortTaskBlock() calls at the end of every test as these are now done in the core code. - FreeRTOS heap back down to 16 kbytes, can probably go lower.  Also added hooks to stack overflow and malloc fail. - switch to using HW timer 1 as that won't conflict with SOFT DEVICE if someone happened to be using it.
BUT there is still debug code in cellular_port_uart.c and things are STILL not stable.  If I run just the first half of cellularPortTestEverything(), where just the OS stuff is tested, nothing fancy, mostly just through-calls into FreeRTOS, then if I comment out the first call to cellularPortLog() it passes, if I leave that call in FreeRTOS complains that the TCB it's trying to delete when destroying the task wasn't allocated in the first place.  Unh?  I've also seen FreeRTOS sit in the idle loop forever when I successfully power off the module, though that may be a separate issue.  I don't think it's FreeRTOS that's at fault; given the evidence above I'm beginning to be suspicious of the NRF_LOG mechanism.  Might try looking at using SWO instead.