STM32F4 portion of the OS changes.
The OS port on ST32F4 now uses the CMSIS-OS API with one exception: there is no exact CMSIS-OS equivalent of our OS queue API.  On the face of it the CMSIS-OS message API does the job, in that under the hood it is mapped by ST to the FreeRTOS queue API, however the size of a message is limited to 32 bits.  There is a CMSIS-OS mail API with flexible content size however that does malloc()/free() for each item which has the potential to fragment the heap all to hell.  The beauty of the FreeRTOS-style queue API is that it is malloc()ed once at start of day and hence doesn't suffer from that problem.

So, rather than (a) be fragmented or (b) limit our OS queue API to 32-bit payloads, in the cas of the OS queue API I have ported it directly to the underlying FreeRTOS queue API rather than going through the constricting CMSIS-OS API.