Fix a design problem in UART handling, which went as follows.
When powering down a module the UART lines can be held low or high by the module, resulting in 0xFF or 0x00 being recognised by our UART as break/overrun or whatever.  Notifications of the arrival of these duff characters can get thrown at the AT URC handler task at quite a rate, filling up the event queue and preventing the AT deinit() function from putting its own "please terminate now" instruction onto the same queue; should never mix data and control, I knew that already.  Anyway, now there is a separate control queue for the AT URC handler task.  This entailed adding a "try receive" call to both the UART and OS queue APIs so that the receiving task can wait for a single event on one, process it or time-out, then try the other.  Both get a look-in and so the lock-out will not occur.  The wait times in the AT URC handler are nice and long for the UART receive and very short for the control queue receive in order not to adversely affect URC reception.

Also:

- Add AT+UPSV=0 into the module configuration stuff at startup, just for good measure.  Like AT+CPSMS=0, this can be removed when we support UART power saving in the driver.
- Add the powerAndAliveness test back in for SARA-R4 since the module has stopped misbehaving in that respect.  We don't know why it was misbehaving: a finger was pointed at AT+UPSV but that is set to 0.  The EVK board the module is mounted on has been power-cycled since I last tried.  Dunno.  Chalk this one up to experience.
- Remove from the Python script the warning that characters on which Unicode decode has failed imply that the target has crashed; they don't, these decode errors seem to happen anyway sometimes at target startup, even though as far as I can tell there are no badly formed ASCII characters being sent from the target.  Some sort of Python/Windows/Unicode issue that I can't do anything about, now ignored.
- Remove the 'x' from the end of the "send all characters" array in cellular_sock_test.c since it's not actually used (sizeof() the array naturally includes the NULL terminator on the end of the array since the array contents are a string).
- Editorial update to NRF52840 SES SDK README.md.